✨ feat: 약관 정보 확인화면 추가
