+++ src/components/user/groups/UserSupportGroup/UserSupportGroup.style.ts
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native';
+
+const userSupportGroupStyles = StyleSheet.create({
+  userControllerContainer: {
+    paddingTop: 20,
+    paddingHorizontal: 20,
+    paddingBottom: 100,
+    gap: 20,
+  },
+});
+
+export default userSupportGroupStyles;
