+++ src/constants/user/userConstants.ts
@@ -26,6 +26,10 @@ const MENT_USER = Object.freeze({
     CUSTOMER_SERVICE_CENTER: '고객센터',
     FAQ: 'FAQ',
     ANNOUNCEMENT: '공지사항',
+    TERM: '약관',
+    TERM_TO_USE: '서비스 이용약관',
+    TERM_TO_PRIVACY: '개인정보 수집',
+    TERM_TO_MARKETING: '마케팅 정보 수신동의',
     INQUIRY: '문의하기',
     SETTING: '설정',
     SERVICE_SETTING: '서비스 설정',
