+++ src/screens/user/UserScreen/UserScreen.tsx
@@ -7,27 +7,17 @@ import UserFieldBoxGroup from 'components/user/groups/UserFieldBoxGroup/UserFiel
 import UserMenuButtonGroup from 'components/user/groups/UserMenuButtonGroup/UserMenuButtonGroup';
 import MENT_USER from 'constants/user/userConstants';
 import fieldData from 'data/lists/fieldData';
-import {
-  Platform,
-  NativeModules,
-  Image,
-  Pressable,
-  ScrollView,
-  TouchableOpacity,
-  View,
-  Linking,
-} from 'react-native';
+import { Image, Pressable, TouchableOpacity, View } from 'react-native';
 import useNavigator from 'hooks/navigator/useNavigator';
 import { useLogout, useMyInfo } from 'hooks/queries/user';
-import useDialog from 'hooks/app/useDialog';
 import WithIconLoading from 'components/suspense/loading/WithIconLoading/WithIconLoading';
 import { colors } from 'styles/theme';
+import UserSupportGroup from 'components/user/groups/UserSupportGroup/UserSupportGroup';
 import userScreenStyles from './UserScreen.style';
 
 const UserScreen = () => {
   const { data: userInfo } = useMyInfo();
   const { stackNavigation } = useNavigator();
-  const { openDialog } = useDialog();
 
   const handleEditProfile = () => {
     stackNavigation.navigate('UserProfileEdit');
@@ -37,43 +27,7 @@ const UserScreen = () => {
     stackNavigation.navigate('UserInterest');
   };
 
-  const handleShowSettingScreen = () => {
-    if (Platform.OS === 'ios') {
-      openSettings();
-    } else {
-      NativeModules.ExternalURLModule.linkAndroidSettings();
-    }
-  };
-
-  const { mutateAsync: logout, isLoading: isLogoutLoading } = useLogout();
-
-  const handleLogout = async () => {
-    openDialog({
-      type: 'warning',
-      text: '로그아웃하시겠습니까?',
-      applyText: '예',
-      closeText: '아니오',
-      apply: async () => {
-        await logout();
-      },
-    });
-  };
-
-  const handleShowFAQ = () => {
-    Linking.openURL(
-      'https://navy-web.notion.site/FAQ-0449b749375646c4b3a55fccc1e44ff7?pvs=4',
-    );
-  };
-
-  const handleShowAnnoincement = () => {
-    Linking.openURL(
-      'https://navy-web.notion.site/9ff02822c01c4a5a905e197389d8b3e4?pvs=4',
-    );
-  };
-
-  const handleShowInquiry = () => {
-    Linking.openURL('http://pf.kakao.com/_QuKlG');
-  };
+  const { isLoading: isLogoutLoading } = useLogout();
 
   return (
     <View style={userScreenStyles.container}>
@@ -146,35 +100,7 @@ const UserScreen = () => {
       </View>
       <Divider height={8} color="darkGrey" />
 
-      <ScrollView
-        showsVerticalScrollIndicator={false}
-        contentContainerStyle={userScreenStyles.userControllerContainer}
-      >
-        <Text variant="bodySB" color="darkGrey">
-          {MENT_USER.MAIN.CUSTOMER_SERVICE_CENTER}
-        </Text>
-        <Pressable onPress={handleShowFAQ}>
-          <Text variant="body2">{MENT_USER.MAIN.FAQ}</Text>
-        </Pressable>
-        <Pressable onPress={handleShowAnnoincement}>
-          <Text variant="body2">{MENT_USER.MAIN.ANNOUNCEMENT}</Text>
-        </Pressable>
-        <Pressable onPress={handleShowInquiry}>
-          <Text variant="body2">{MENT_USER.MAIN.INQUIRY}</Text>
-        </Pressable>
-        <Divider height={1} color="darkGrey" />
-        <Pressable>
-          <Text variant="bodySB" color="darkGrey">
-            {MENT_USER.MAIN.SETTING}
-          </Text>
-        </Pressable>
-        <Pressable onPress={handleShowSettingScreen}>
-          <Text variant="body2">{MENT_USER.MAIN.SERVICE_SETTING}</Text>
-        </Pressable>
-        <Pressable onPress={handleLogout}>
-          <Text variant="body2">{MENT_USER.MAIN.LOGOUT}</Text>
-        </Pressable>
-      </ScrollView>
+      <UserSupportGroup />
     </View>
   );
 };
